Added option to change base command and aliases
