+++ src/main/resources/config.toml
@@ -11,3 +11,13 @@ uuidupdated = true
 
 # Cooldown between placing protection blocks (in seconds). -1 to disable.
 placing_cooldown = -1
+
+# Base command for protection stones (change if conflicting with other commands)
+base_command = "ps"
+
+# Aliases for the command
+aliases = [
+    "pstone",
+    "protectionstone",
+    "protectionstones"
+]
